Add 'models' folder, 'utils' folder and 'simulation.py' for better encapsulation
The 'models' folder contain elements of our ridesharing problem that needs to be modelled. Currently there is the Environment model and Passenger model. To better encapsulate them, the Passenger class is removed from 'environment.py'.

The 'utils' folder is added to avoid cluttering the model classes, as well as providing frequently re-usable functions. The function that produces the string representation of an 'Environment' instance is moved to utils/info_utils.py.

The 'simulation.py' file will be responsible for running simulation scripts.