Add 'identifier' attribute to each vertex
This additional attribute is required because the underlying index of each vertex represented by igraph will change as the graph structure changes, or if a subgraph is produced, while the attributes will not change. Getting a vertex from an igraph object requires knowing the underlying index. In order to make things simpler, new functions were created under 'graph_utils.py' that queries for a specific vertex object based on its 'identifier' attribute. This means that we do not need to worry and keep track of the underlying index of a vertex. Throughout the codebase, 'id' refers to the 'identifier' attribute while 'index' refers to the underlying index of the vertex.