Use list to store agents instead of set for greedy insert
Set data structure introduced randomness in terms of the ordering of passengers for insertion. To ensure reproducable results, the seed for the greedy insert algorithm is first set. Next, numpy.shuffle is used to shuffle the ordering of agents before each iteration. Lastly, the first agent to be inserted will be the first agent from the shuffled list